+++ controllers/player.js
@@ -1,9 +1,23 @@
 const playerRouter = require('express').Router()
 const Player = require('../models/player')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = (req) => {
+  const authorization = req.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+}
 
 playerRouter.post('/', async (req, res) => {
   const {name, elo} = req.body
 
+  const decodedToken = jwt.verify(getTokenFrom(req), process.env.SECRET)                         
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: 'token invalid' })
+  }
+  
   const player = new Player({
     name: name,
     elo: elo
@@ -43,6 +57,11 @@ playerRouter.get('/:id', async (req, res) => {
 })
 
 playerRouter.delete('/:id', async (req, res) => {
+  const decodedToken = jwt.verify(getTokenFrom(req), process.env.SECRET)                         
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: 'token invalid' })
+  }
+  
   await Player.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
